refactor: extract useModelImportHandlers + ImportProgressModal (#3501 Phase 1k)
